fix: Make PortalsPubSub.publish non-blocking (#25)
fix: The publish function was blocking callers who publish until all subscribers were published to. In practice this is unlikely to be
noticeable, but if there are many subscribers it could potentially block
the caller for an inordinate amount of time.
chore: Add test helpers for Combine and update tests to use them.
chore: Remove NSLog from objective-c tests.